refactor: Moved file centric view and serializer to file execution app
