+++ backend/workflow_manager/execution/serializer/__init__.py
@@ -1,2 +1,1 @@
 from .execution import ExecutionSerializer  # noqa: F401
-from .file_centric import FileCentricExecutionSerializer  # noqa: F401
