+++ backend/workflow_manager/execution/views/__init__.py
@@ -1,2 +1,1 @@
 from .execution import ExecutionViewSet  # noqa: F401
-from .file_centric import FileCentricExecutionViewSet  # noqa: F401
